% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parallelShowOptions.R
\name{parallelGetOptions}
\alias{parallelGetOptions}
\title{Displays the configured package options.}
\usage{
parallelGetOptions()
}
\description{
Displayed are current and default settings.

For details on the configuration procedure please read
\code{\link{parallelStart}} and \url{https://github.com/berndbischl/parallelMap}.
}

