% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallelRegisterLevels.R
\name{parallelGetRegisteredLevels}
\alias{parallelGetRegisteredLevels}
\title{Get registered parallelization levels for all currently loaded packages.}
\usage{
parallelGetRegisteredLevels()
}
\value{
Returns a named list with all registered levels.
}
\description{
Get registered parallelization levels for all currently loaded packages.
}

