\name{parallelShowOptions}
\alias{parallelShowOptions}
\title{Displays the configured package options.}
\usage{
parallelShowOptions()
}
\description{
Displayed are current and default settings.
}
\details{
For details on the configuration procedure please read
\code{\link{parallelStart}} and
\url{https://github.com/berndbischl/parallelMap}.
}

