% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parallelRegisterLevels.R
\name{parallelGetRegisteredLevels}
\alias{parallelGetRegisteredLevels}
\title{Get registered parallelization levels for all currently loaded packages.}
\usage{
parallelGetRegisteredLevels()
}
\value{
[code{RegisteredLevels}]. See above.
}
\description{
The S3 object only has one slot, which is called \code{levels}.
This contains a named list. Each name refers to \code{package} from the call to
\code{\link{parallelRegisterLevels}}, while the entries are character
vectors of the form \dQuote{package.level}.
}

