% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parallelRegisterLevels}
\alias{parallelRegisterLevels}
\title{Register available parallelization levels in a client package.}
\usage{
parallelRegisterLevels(package, levels)
}
\arguments{
\item{package}{[\code{character(1)}]\cr
Name of your package.}

\item{levels}{[\code{character(1)}]\cr
Availabe levels that are used in the \code{\link{parallelMap}} operations of your package.}
}
\value{
Nothing.
}
\description{
Call this in your\code{\link{.onLoad}} in your \dQuote{zzz.R}.
}

