% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parallelShowOptions.R
\name{parallelShowOptions}
\alias{parallelShowOptions}
\title{Displays the configured package options.}
\usage{
parallelShowOptions()
}
\description{
Displayed are current and default settings.
}
\details{
For details on the configuration procedure please read
\code{\link{parallelStart}} and \url{https://github.com/berndbischl/parallelMap}.
}

