% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallelRegisterLevels.R
\name{parallelShowRegisteredLevels}
\alias{parallelShowRegisteredLevels}
\title{Display registered parallelization levels for all currently loaded packages.}
\usage{
parallelShowRegisteredLevels()
}
\value{
Invisibly returns a list object that contains the displayed information.
}
\description{
Display registered parallelization levels for all currently loaded packages.
}

