% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelSource.R
\name{parallelSource}
\alias{parallelSource}
\title{Source R files for parallelization.}
\usage{
parallelSource(..., files, master = TRUE, level = NA_character_,
  show.info = NA)
}
\arguments{
\item{...}{\link{character}\cr
File paths to sources.}

\item{files}{\link{character}\cr
File paths to sources.
Alternative way to pass arguments.}

\item{master}{(\code{logical(1)})\cr
Source files also on master for any mode?
Default is \code{TRUE}.}

\item{level}{(\code{character(1)})\cr
If a (non-missing) level is specified in \code{\link[=parallelStart]{parallelStart()}},
the function only sources the files if the level specified here matches.
See \code{\link[=parallelMap]{parallelMap()}}.
Useful if this function is used in a package.
Default is \code{NA}.}

\item{show.info}{(\code{logical(1)})\cr
Verbose output on console?
Can be used to override setting from options / \code{\link[=parallelStart]{parallelStart()}}.
Default is NA which means no overriding.}
}
\value{
Nothing.
}
\description{
Makes sure that the files are sourced in slave process so that
they can be used in a job function which is later run with \code{\link[=parallelMap]{parallelMap()}}.

For all modes, the files are also (potentially) loaded on the master.
}
